Before error

+++ D11/program.cpp
@@ -19,7 +19,7 @@ Cqueue::~Cqueue()
 void Cqueue::Ielikt(char elements)
 {
     for(int i = 0; i < 5; ++i){
-        cout << rinda[i] << "|";
+        cout << (int)rinda[i] << "|";
     }
     cout << endl;
     if (garums == 5)
@@ -31,7 +31,7 @@ void Cqueue::Ielikt(char elements)
     rinda[pozicija] = elements;
 
     for(int i = 0; i < 5; ++i){
-        cout << rinda[i] << "|";
+        cout << (int)rinda[i] << "|";
     }
     cout << endl;
